Extract StoryContainer
