+++ stories/components/StoryContainer.js
@@ -1,0 +1,33 @@
+/**
+ * Provides global style primitives so that comonents rendered in the storybook
+ * look like those rendered in the app.
+ *
+ * @flow
+ */
+import type {Node} from 'react';
+
+import React, {Component} from 'react';
+import styled from 'react-emotion';
+
+import injectGlobals from 'universal/styles/hepha';
+import globalStyles from 'universal/styles/theme/globalStyles';
+
+type Props = {
+  render: () => Node
+};
+
+const FullPageWrapper = styled('div')({
+  height: '100vh',
+  padding: '1rem',
+  width: '100vw'
+});
+
+export default class StoryContainer extends Component<Props> {
+  componentWillMount() {
+    injectGlobals(globalStyles);
+  }
+
+  render() {
+    return <FullPageWrapper>{this.props.render()}</FullPageWrapper>;
+  }
+}
